+++ api/apidocs.go
@@ -1,0 +1,8 @@
+package apidocs
+
+import _ "embed"
+
+//go:embed openapi.json
+var JSONOpenAPISpecs string
+//go:embed openapi.yaml
+var YAMLOpenAPISpecs string