+++ api/convert.sh
@@ -1,0 +1,1 @@
+yq . openapi.yaml > openapi.json